+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+
 ## [v1.3.0] - 2016-09-??
 ### Added
 - Scripts: